pyproject.toml
[project]
name = "mcp-purecloud-social-listening"
version = "0.1.0"
description = "MCP Server with FastAPI and OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.21",
    "pydantic>=2.10.4",
    "mcp[cli]>=1.25.0",
    "openai-agents>=0.6.7",
    "python-dotenv>=1.2.1",
    "purecloudplatformclientv2>=248.0.0",
]
